+++ plugins/devel/eventExample.cpp
@@ -128,26 +128,26 @@ void jobCompleted(color_ostream& out, void* job) {
 }
 
 void timePassed(color_ostream& out, void* ptr) {
-    out.print("Time: %d\n", (int32_t)(ptr));
+    out.print("Time: %d\n", (intptr_t)(ptr));
 }
 
 void unitDeath(color_ostream& out, void* ptr) {
-    out.print("Death: %d\n", (int32_t)(ptr));
+    out.print("Death: %d\n", (intptr_t)(ptr));
 }
 
 void itemCreate(color_ostream& out, void* ptr) {
-    int32_t item_index = df::item::binsearch_index(df::global::world->items.all, (int32_t)ptr);
+    int32_t item_index = df::item::binsearch_index(df::global::world->items.all, (intptr_t)ptr);
     if ( item_index == -1 ) {
         out.print("%s, %d: Error.\n", __FILE__, __LINE__);
     }
     df::item* item = df::global::world->items.all[item_index];
     df::item_type type = item->getType();
     df::coord pos = item->pos;
-    out.print("Item created: %d, %s, at (%d,%d,%d)\n", (int32_t)(ptr), ENUM_KEY_STR(item_type, type).c_str(), pos.x, pos.y, pos.z);
+    out.print("Item created: %d, %s, at (%d,%d,%d)\n", (intptr_t)(ptr), ENUM_KEY_STR(item_type, type).c_str(), pos.x, pos.y, pos.z);
 }
 
 void building(color_ostream& out, void* ptr) {
-    out.print("Building created/destroyed: %d\n", (int32_t)ptr);
+    out.print("Building created/destroyed: %d\n", (intptr_t)ptr);
 }
 
 void construction(color_ostream& out, void* ptr) {
@@ -168,7 +168,7 @@ void syndrome(color_ostream& out, void* ptr) {
 }
 
 void invasion(color_ostream& out, void* ptr) {
-    out.print("New invasion! %d\n", (int32_t)ptr);
+    out.print("New invasion! %d\n", (intptr_t)ptr);
 }
 
 void unitAttack(color_ostream& out, void* ptr) {
